requirements.txt
Django==4.0.2
djangorestframework==3.13.1
django-filter==21.1
simplejson==3.17.6
requests==2.27.1
django-apscheduler==0.6.0
python-telegram-bot==13.11
PyMySQL >= 1.0.2
lxml >= 4.7.1
beautifulsoup4 >= 4.10.0
ddddocr >= 1.4.3
#####
#gunicorn

#django-unixtimestampfield
#gevent

